+++ lib/triggers/getOrganizationsPolling.js
@@ -18,10 +18,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-const Q = require('q');
-const request = require('request-promise');
-const { messages } = require('elasticio-node');
-const { createSession } = require('./../utils/wice');
+const { newMessage, getOrganizations } = require('./../utils/helpers');
 
 /**
  * This method will be called from elastic.io platform providing following data
@@ -30,73 +27,16 @@ const { createSession } = require('./../utils/wice');
  * @param cfg configuration that is account information and configuration field values
  * @param snapshot saves the current state of integration step for the future reference
  */
-function processTrigger(msg, cfg, snapshot = {}) {
-  const { applicationUid, domainId, schema } = cfg;
-  const self = this;
-  let organizations = [];
+async function processTrigger(msg, cfg, snapshot = {}) {
+  try {
+    const { applicationUid, domainId, schema } = cfg;
 
-  // Set the snapshot if it is not provided
-  snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).toISOString();
-  console.log(`Last Updated: ${snapshot.lastUpdated}`);
+    // Set the snapshot if it is not provided
+    snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).toISOString();
+    console.log(`Last Updated: ${snapshot.lastUpdated}`);
 
-  function customOrganization(organization) {
-    const customOrganizaiontFormat = {
-      rowid: organization.rowid,
-      last_update: organization.last_update,
-      name: organization.name,
-      email: organization.email,
-      phone: organization.phone,
-      fax: organization.fax,
-      street: organization.street,
-      street_number: organization.street_number,
-      zip_code: organization.zip_code,
-      p_o_box: organization.p_o_box,
-      town: organization.town,
-      state: organization.state,
-      country: organization.country,
-    };
-    return customOrganizaiontFormat;
-  }
+    const organizations = await getOrganizations(cfg, snapshot);
 
-  async function fetchAll(options) {
-    try {
-      const result = [];
-      const organizationsResponse = await request.get(options);
-      const organizationsObj = JSON.parse(organizationsResponse);
-
-      if (organizationsObj.loop_addresses === undefined) return result;
-
-      organizationsObj.loop_addresses.filter((organization) => {
-        const currentOrganization = customOrganization(organization);
-        currentOrganization.last_update > snapshot.lastUpdated && result.push(currentOrganization);
-      });
-
-      result.sort((a, b) => Date.parse(a.last_update) - Date.parse(b.last_update));
-      return result;
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-
-
-  async function getOrganizations() {
-    try {
-      const cookie = await createSession(cfg);
-      const options = {
-        uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json?method=get_all_companies&full_list=1&cookie=${cookie}`,
-        headers: { 'X-API-KEY': cfg.apikey },
-      };
-      organizations = await fetchAll(options);
-
-      if (!organizations || !Array.isArray(organizations)) return `Expected records array. Instead received: ${JSON.stringify(organizations)}`;
-
-      return organizations;
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-
-  async function emitData() {
     /** The following block creates the meta object.
      *  This meta object stores information which are later needed in order to make the hub and spoke architecture work properly
      */
@@ -115,31 +55,18 @@ function processTrigger(msg, cfg, snapshot = {}) {
         delete elem.rowid;
         newElement.meta = oihMeta;
         newElement.data = elem;
-        self.emit('data', messages.newMessageWithBody(newElement));
+        this.emit('data', newMessage(newElement));
       });
       snapshot.lastUpdated = organizations[organizations.length - 1].last_update;
       console.log(`New snapshot: ${snapshot.lastUpdated}`);
-      self.emit('snapshot', snapshot);
+      this.emit('snapshot', snapshot);
     } else {
-      self.emit('snapshot', snapshot);
+      this.emit('snapshot', snapshot);
     }
-  }
-
-  function emitError(e) {
+  } catch (e) {
     console.log(`ERROR: ${e}`);
-    self.emit('error', e);
+    this.emit('error', e);
   }
-
-  function emitEnd() {
-    console.log('Finished execution');
-    self.emit('end');
-  }
-
-  Q()
-    .then(getOrganizations)
-    .then(emitData)
-    .fail(emitError)
-    .done(emitEnd);
 }
 
 exports.process = processTrigger;
