+++ lib/triggers/getPersonsPolling.js
@@ -18,10 +18,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
  */
 
-const Q = require('q');
-const request = require('request-promise');
-const { messages } = require('elasticio-node');
-const { createSession } = require('./../utils/wice');
+const { getPersons, newMessage } = require('./../utils/helpers');
 
 /**
  * This method will be called from elastic.io platform providing following data
@@ -30,81 +27,15 @@ const { createSession } = require('./../utils/wice');
  * @param cfg configuration that is account information and configuration field values
  * @param snapshot saves the current state of integration step for the future reference
  */
-function processTrigger(msg, cfg, snapshot = {}) {
-  const { applicationUid, domainId, schema } = cfg;
-  const self = this;
-  let contacts = [];
+async function processTrigger(msg, cfg, snapshot = {}) {
+  try {
+    const { applicationUid, domainId, schema } = cfg;
 
-  // Set the snapshot if it is not provided
-  snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).toISOString();
-  console.log(`Last Updated: ${snapshot.lastUpdated}`);
+    // Set the snapshot if it is not provided
+    snapshot.lastUpdated = snapshot.lastUpdated || (new Date(0)).toISOString();
+    console.log(`Last Updated: ${snapshot.lastUpdated}`);
 
-  function customPerson(person) {
-    const customUserFormat = {
-      rowid: person.rowid,
-      for_rowid: person.for_rowid,
-      same_contactperson: person.same_contactperson,
-      last_update: person.last_update,
-      deactivated: person.deactivated,
-      name: person.name,
-      firstname: person.firstname,
-      email: person.email,
-      title: person.title,
-      salutation: person.salutation,
-      birthday: person.birthday,
-      private_street: person.private_street,
-      private_street_number: person.private_street_number,
-      private_zip_code: person.private_zip_code,
-      private_town: person.private_town,
-      private_state: person.private_state,
-      private_country: person.private_country,
-      phone: person.phone,
-      fax: person.fax,
-      private_phone: person.private_phone,
-      private_mobile_phone: person.private_mobile_phone,
-      private_email: person.private_email,
-    };
-    return customUserFormat;
-  }
-
-  async function fetchAll(options) {
-    try {
-      const result = [];
-      const persons = await request.get(options);
-      const personsObj = JSON.parse(persons);
-
-      if (personsObj.loop_addresses === undefined) return result;
-
-      personsObj.loop_addresses.filter((person) => {
-        const currentPerson = customPerson(person);
-        currentPerson.last_update > snapshot.lastUpdated && result.push(currentPerson);
-      });
-
-      result.sort((a, b) => Date.parse(a.last_update) - Date.parse(b.last_update));
-      return result;
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-
-  async function getPersons() {
-    try {
-      const cookie = await createSession(cfg);
-      const options = {
-        uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json?method=get_all_persons&full_list=1&cookie=${cookie}`,
-        headers: { 'X-API-KEY': cfg.apikey },
-      };
-
-      contacts = await fetchAll(options);
-      if (!contacts || !Array.isArray(contacts)) return `Expected records array. Instead received: ${JSON.stringify(contacts)}`;
-
-      return contacts;
-    } catch (e) {
-      throw new Error(e);
-    }
-  }
-
-  async function emitData() {
+    const contacts = await getPersons(cfg, snapshot);
     /** The following block creates the meta object.
      *  This meta object stores information which are later needed in order to make the hub and spoke architecture work properly
      */
@@ -123,31 +54,18 @@ function processTrigger(msg, cfg, snapshot = {}) {
         delete elem.rowid;
         newElement.meta = oihMeta;
         newElement.data = elem;
-        self.emit('data', messages.newMessageWithBody(newElement));
+        this.emit('data', newMessage(newElement));
       });
       snapshot.lastUpdated = contacts[contacts.length - 1].last_update;
       console.log(`New snapshot: ${snapshot.lastUpdated}`);
-      self.emit('snapshot', snapshot);
+      this.emit('snapshot', snapshot);
     } else {
-      self.emit('snapshot', snapshot);
+      this.emit('snapshot', snapshot);
     }
-  }
-
-  function emitError(e) {
+  } catch (e) {
     console.log(`ERROR: ${e}`);
-    self.emit('error', e);
+    this.emit('error', e);
   }
-
-  function emitEnd() {
-    console.log('Finished execution');
-    self.emit('end');
-  }
-
-  Q()
-    .then(getPersons)
-    .then(emitData)
-    .fail(emitError)
-    .done(emitEnd);
 }
 
 exports.process = processTrigger;
