+++ lib/utils/helpers.js
@@ -26,6 +26,7 @@ async function createSession(cfg) {
     console.log(`ERROR: ${e}`);
     throw new Error(e);
   }
+}
 
 
 function newMessage(body) {
@@ -40,7 +41,111 @@ function newMessage(body) {
   return msg;
 }
 
+function customPerson(person) {
+  const customUserFormat = {
+    rowid: person.rowid,
+    for_rowid: person.for_rowid,
+    same_contactperson: person.same_contactperson,
+    last_update: person.last_update,
+    deactivated: person.deactivated,
+    name: person.name,
+    firstname: person.firstname,
+    email: person.email,
+    title: person.title,
+    salutation: person.salutation,
+    birthday: person.birthday,
+    private_street: person.private_street,
+    private_street_number: person.private_street_number,
+    private_zip_code: person.private_zip_code,
+    private_town: person.private_town,
+    private_state: person.private_state,
+    private_country: person.private_country,
+    phone: person.phone,
+    fax: person.fax,
+    private_phone: person.private_phone,
+    private_mobile_phone: person.private_mobile_phone,
+    private_email: person.private_email,
+  };
+  return customUserFormat;
+}
+
+function customOrganization(organization) {
+  const customOrganizaiontFormat = {
+    rowid: organization.rowid,
+    last_update: organization.last_update,
+    name: organization.name,
+    email: organization.email,
+    phone: organization.phone,
+    fax: organization.fax,
+    street: organization.street,
+    street_number: organization.street_number,
+    zip_code: organization.zip_code,
+    p_o_box: organization.p_o_box,
+    town: organization.town,
+    state: organization.state,
+    country: organization.country,
+  };
+  return customOrganizaiontFormat;
+}
+
+async function fetchAll(options, mapping, snapshot) {
+  try {
+    const result = [];
+    const response = await request.get(options);
+    const responseObj = JSON.parse(response);
+
+    if (responseObj.loop_addresses === undefined) return result;
+
+    responseObj.loop_addresses.forEach((object) => {
+      const currentObject = mapping(object);
+      if (currentObject.last_update > snapshot.lastUpdated) result.push(currentObject);
+    });
+
+    result.sort((a, b) => Date.parse(a.last_update) - Date.parse(b.last_update));
+    return result;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+
+async function getPersons(cfg, snapshot) {
+  try {
+    const cookie = await createSession(cfg);
+    const options = {
+      method: 'get',
+      uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json?method=get_all_persons&full_list=1&cookie=${cookie}`,
+      headers: { 'X-API-KEY': cfg.apikey },
+    };
+
+    const contacts = await fetchAll(options, customPerson, snapshot);
+    if (!contacts || !Array.isArray(contacts)) return `Expected records array. Instead received: ${JSON.stringify(contacts)}`;
+
+    return contacts;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
+async function getOrganizations(cfg, snapshot) {
+  try {
+    const cookie = await createSession(cfg);
+    const options = {
+      uri: `${cfg.url}/plugin/wp_wice_client_api_backend/json?method=get_all_companies&full_list=1&cookie=${cookie}`,
+      headers: { 'X-API-KEY': cfg.apikey },
+    };
+
+    const organizations = await fetchAll(options, customOrganization, snapshot);
+
+    if (!organizations || !Array.isArray(organizations)) return `Expected records array. Instead received: ${JSON.stringify(organizations)}`;
+
+    return organizations;
+  } catch (e) {
+    throw new Error(e);
+  }
+}
+
 
 module.exports = {
-  createSession, newMessage,
+  createSession, newMessage, getPersons, getOrganizations,
 };
